+++ src/main/java/com/dzui/shoesshop/repositories/GenderRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.dzui.shoesshop.entities.Gender;
 
-public interface GenderRepository extends CrudRepository<Gender, Long> {
+public interface GenderRepository extends CrudRepository<Gender, Integer> {
 
 }
